move isolated functions into separate files
